requirements.txt
# Core Framework
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0

# Vector Database
faiss-cpu>=1.8.0

# UI & Environment
streamlit>=1.35.0
python-dotenv>=1.0.0

# LLM Clients
openai>=1.30.0
google-generativeai>=0.7.0

# Utilities
sentence-transformers>=3.0.0
requests>=2.32.0
packaging>=24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
